NewBreeze v3.0.0. Removed QStorageInfo dependency of NBDeviceInfo.

+++ app/Gui/Widgets/SidePanel/NBDeviceView.hpp
@@ -112,7 +112,7 @@ class NBDeviceAction : public QWidget {
 
 	private:
 		QString mMountPoint;
-		quint64 percentUsed;
+		qreal percentUsed;
 		QIcon icon;
 		QString mDeviceLabel;
